place all user data in the dashboard provider
