+++ src/app/(All Page Here)/user-dashboard/[userId]/page.jsx
@@ -15,51 +15,7 @@ import React, { use, useEffect, useState } from 'react';
 
 const UserDashboard = () => {
 
-    const [appointmentData, setAppointmentData] = useState([])
 
-    const [userHealthStats, setHealthStats] = useState([])
-
-    const [userToDo, setUserToDo] = useState([])
-
-
-    const [userGoals, setUserGoal] = useState()
-
-
-    const { userId } = useParams()
-
-    const { user } = use(AuthContext)
-
-
-
-
-
-    useEffect(() => {
-        const getUserStats = async () => {
-
-            try {
-                // url
-                const healthStatsUrl = `/api/get-health-stats/${userId}`;
-                const doctorListUrl = `/api/get-appointment-list/${userId}`;
-                const toDoUrl = `/api/get-todo-task/${userId}`;
-
-                // response
-                const healthStatsResponse = await axios.get(healthStatsUrl)
-                const doctorListResponse = await axios.get(doctorListUrl)
-                const toDoListResponse = await axios.get(toDoUrl)
-
-                // set data in state
-                setHealthStats(healthStatsResponse?.data[0]?.userStats)
-                setAppointmentData(doctorListResponse?.data[0]?.appointmentDetails)
-                setUserToDo(toDoListResponse?.data[0]?.todo)
-
-            } catch (error) {
-                console.error('Error fetching health stats:', error);
-            }
-        }
-
-        getUserStats()
-
-    }, [userId])
 
     return (
